# enhance search performance