attempt to enhance search performance by branchless binary search.
concrete implementation of sort.Search is fastest of the candidates.

https://en.algorithmica.org/hpc/data-structures/binary-search/

bool to int conversion without jump: https://0x0f.me/blog/golang-compiler-optimization/

```
BenchmarkBinarySearchForPosting

goos: linux
goarch: amd64
pkg: github.com/getumen/sakuin/posting
cpu: Intel(R) Core(TM) i7-7600U CPU @ 2.80GHz
BenchmarkBinarySearchForPosting
BenchmarkBinarySearchForPosting/sort.Slice
BenchmarkBinarySearchForPosting/sort.Slice-4         	25396570	        48.46 ns/op	       0 B/op	       0 allocs/op
BenchmarkBinarySearchForPosting/standard
BenchmarkBinarySearchForPosting/standard-4           	62686153	        19.53 ns/op	       0 B/op	       0 allocs/op
BenchmarkBinarySearchForPosting/branch-prediction
BenchmarkBinarySearchForPosting/branch-prediction-4  	44398458	        24.13 ns/op	       0 B/op	       0 allocs/op
BenchmarkBinarySearchForPosting/branchless
BenchmarkBinarySearchForPosting/branchless-4         	15233991	        69.40 ns/op	       0 B/op	       0 allocs/op
PASS
```

```
BenchmarkBinarySearchForPosition

goos: linux
goarch: amd64
pkg: github.com/getumen/sakuin/posting
cpu: Intel(R) Core(TM) i7-7600U CPU @ 2.80GHz
BenchmarkBinarySearchForPosition
BenchmarkBinarySearchForPosition/sort.Slice
BenchmarkBinarySearchForPosition/sort.Slice-4         	26721198	        42.52 ns/op	       0 B/op	       0 allocs/op
BenchmarkBinarySearchForPosition/standard
BenchmarkBinarySearchForPosition/standard-4           	69359968	        17.25 ns/op	       0 B/op	       0 allocs/op
BenchmarkBinarySearchForPosition/branch-prediction
BenchmarkBinarySearchForPosition/branch-prediction-4  	47127693	        22.64 ns/op	       0 B/op	       0 allocs/op
BenchmarkBinarySearchForPosition/branchless
BenchmarkBinarySearchForPosition/branchless-4         	23144769	        46.94 ns/op	       0 B/op	       0 allocs/op
PASS
```